download-fits/requirements.txt
google-cloud-pubsub
aiostream
asyncio
aiohttp
aiofiles
